Chore(.feign): move client, dto
