+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/client/KakaoAuthorizeFeignClient.java
@@ -1,11 +1,11 @@
-package com.newbarams.ajaja.infra.feign.kakao;
+package com.newbarams.ajaja.infra.feign.kakao.client;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
-import com.newbarams.ajaja.module.user.kakao.model.KakaoTokenRequest;
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoResponse;
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoTokenRequest;
 
 @FeignClient(name = "KakaoAuthorizeFeignClient", url = "https://kauth.kakao.com")
 public interface KakaoAuthorizeFeignClient {
