+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/client/KakaoProfileFeignClient.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.infra.feign.kakao;
+package com.newbarams.ajaja.infra.feign.kakao.client;
 
 import static org.springframework.http.HttpHeaders.*;
 
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import com.newbarams.ajaja.module.user.kakao.model.KakaoResponse;
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoResponse;
 
 @FeignClient(name = "KakaoProfileFeignClient", url = "https://kapi.kakao.com")
 public interface KakaoProfileFeignClient {
