+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/client/KakaoUnlinkFeignClient.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.infra.feign.kakao;
+package com.newbarams.ajaja.infra.feign.kakao.client;
 
 import static org.springframework.http.HttpHeaders.*;
 
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 
-import com.newbarams.ajaja.module.user.kakao.model.KakaoUnlinkRequest;
+import com.newbarams.ajaja.infra.feign.kakao.model.KakaoUnlinkRequest;
 
 @FeignClient(name = "KakaoDisconnectFeignClient", url = "https://kapi.kakao.com")
 public interface KakaoUnlinkFeignClient {
