+++ src/main/java/com/newbarams/ajaja/infra/feign/kakao/model/KakaoResponse.java
@@ -1,10 +1,10 @@
-package com.newbarams.ajaja.module.user.kakao.model;
+package com.newbarams.ajaja.infra.feign.kakao.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.newbarams.ajaja.module.user.application.model.AccessToken;
-import com.newbarams.ajaja.module.user.application.model.Profile;
+import com.newbarams.ajaja.module.auth.application.model.AccessToken;
+import com.newbarams.ajaja.module.auth.application.model.Profile;
 
 import lombok.Data;
